src/animal_detection.py
import cv2
import torch
from ultralytics import YOLO
import serial
import warnings

import smtplib 
import os
import time
import numpy as np 
import requests  # To send the photo to Telegram 
from email.mime.text import MIMEText 
from email.mime.multipart import MIMEMultipart 
from email.mime.image import MIMEImage 
from datetime import datetime

warnings.filterwarnings("ignore")

ser = None
print("Running in simulation mode (no Arduino)")
 # Adjust COM port and baudrate as needed
print("Serial connection opened successfully!")

# Load the YOLOv8 model
model = YOLO("yolov8n.pt") # replace with your model path

# Email credentials 
sender_email = " " #Sender's Email Address
receiver_emails = [" "] #Reciever's Email Address
sender_password = " " #App Password  
smtp_server = "smtp.gmail.com" 
smtp_port = 587

def capture_photo(frame): 
    print("Capturing image from the current frame...") 
 
    timestamp = datetime.now().strftime("%Y-%m-%d_%H-%M-%S") 
    filename = f"photo_{timestamp}.jpg" 
    cv2.imwrite(filename, frame)  # Save the current frame as the image 
    print(f"Photo saved as {filename}") 
     
    return filename 

# Function to send email with the person's name in the subject and body 
def send_mail(person_name, attachment=None): 
    print(f"Sending mail...") 
    time.sleep(2) 
 
    msg = MIMEMultipart() 
    msg['From'] = sender_email 
    msg['To'] = ", ".join(receiver_emails) 
    msg['Subject'] = f"{person_name} detected" 
 
    body_text = f"{person_name} detected at {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}.\n\nPlease find the attached photo." 
    msg.attach(MIMEText(body_text, 'plain')) 
 
    if attachment: 
        with open(attachment, 'rb') as fp: 
            img_data = MIMEImage(fp.read()) 
            msg.attach(img_data) 
 
    with smtplib.SMTP(smtp_server, smtp_port) as server: 
        server.starttls() 
        server.login(sender_email, sender_password) 
        server.sendmail(sender_email, receiver_emails, msg.as_string()) 
 
    print("Mail sent") 
    time.sleep(2) 

# Open webcam (0 for default camera)
cap = cv2.VideoCapture(0)

# List of animal classes to detect
#animal_classes = ["dog", "cat", "cow", "horse", "sheep", "elephant", "bear", "zebra", "giraffe", "bird"]
animal_classes = ["person", "dog", "cat", "cow"]

while cap.isOpened():
    ret, frame = cap.read()
    if not ret:
        break

    # Run YOLOv8 inference on the frame
    results = model(frame)
    results = model(frame, verbose=False)  # Suppressing speed metrics output


    for result in results:
        for box in result.boxes:
            cls_id = int(box.cls[0])  # Class ID
            label = model.names[cls_id]  # Class name
            confidence = float(box.conf[0])  # Confidence score
            
            # Get bounding box coordinates
            x1, y1, x2, y2 = map(int, box.xyxy[0])
            # Draw bounding box and label
            cv2.rectangle(frame, (x1, y1), (x2, y2), (0, 255, 0), 2)
            cv2.putText(frame, f"{label} {confidence:.2f}", (x1, y1 - 10), 
                        cv2.FONT_HERSHEY_SIMPLEX, 0.6, (0, 255, 0), 2)

            # Check if the detected object is an animal and confidence is > 0.7
            if label in animal_classes and confidence > 0.7:
                
                print(f"Detected: {label} with confidence {confidence:.2f}")
                values_string = f"${label}#\n"
                print(f"[SIMULATED TX] {values_string}")
                print(f"data sent : {values_string}")
                photo_filename = capture_photo(frame)  # Capture photo
                 # send_mail(f"{label} detected", photo_filename)
                time.sleep(10)
                

                # Get bounding box coordinates
                x1, y1, x2, y2 = map(int, box.xyxy[0])

                # Draw bounding box and label
                cv2.rectangle(frame, (x1, y1), (x2, y2), (0, 255, 0), 2)
                cv2.putText(frame, f"{label} {confidence:.2f}", (x1, y1 - 10), 
                            cv2.FONT_HERSHEY_SIMPLEX, 0.6, (0, 255, 0), 2)

    # Display the frame
    cv2.imshow("YOLOv8 Animal Detection", frame)

    # Press 'q' to exit
    if cv2.waitKey(1) & 0xFF == ord('q'):
        break

# Release resources
cap.release()
cv2.destroyAllWindows()

