requirements.txt
ultralytics
opencv-python
numpy
torch
pyserial